Push return value in the caller for JIT to JIT call
For JIT to JIT call, we used to push the return value for the caller
in the leave instruction for the callee. This changes it so that the
callee leaves the return value in RAX and the caller pushes it.

For the case where we are called from the interpreter, we save a few
instructions in gen_leave_exit(). For JIT to JIT call, we save two
instructions for adjusting the caller's REG_SP.

The down side is that the code layout is changed. Before this change,
the code for the callee in a JIT to JIT call used to follow the code for
the send. I'll need to benchmark this to see what difference it makes.

Come to think of it, we don't need to make jit_return go to the same
block to get the instruction savings in both type of leaves...