Stamp out some out of memory crashes
This change makes it so that we don't crash with rb_bug() when we run
out of executable memory while generating code for a new entry point
or when hitting a lazy stub.

A new API, `YJIT.simulate_out_of_mem!`, is used to test for out of
memory conditions.

If we run out while generating a new entry point, we can handle it by
not installing an entry point. Handling running out while compiling a
lazy stub is more involved.

An exit for the stub target is added to the the piece of code that calls
branch_stub_hit() and used when branch_stub_hit() returns NULL. Unlike
before, being able to generate new code is no longer required for
lazy stubs to function correctly.

A situation not handled by this change is when we need to invalidate a
block while we are out of memory. We generate new stubs when we
invalidate and that is not handled for now.