Call instructions for the new backend
This adds OP_CALL and OP_CCALL. OP_CALL is just a wrapper around the x86 call instruction, and accepts as its only operand a pointer to the function that should be called. OP_CCALL is a higher-level instruction that accepts a variable number of arguments that should be passed to the function. When allocating registers, OP_CCALL breaks down to a mov and an OP_CALL. OP_CALL is responsible for putting values into the correct registers in order to handle argument order correctly since that's platform-specific.