OP_SELECT_*
This commit adds 9 new instructions to the backend to accomplish selecting a value based on a comparison. The higher-level instructions that it introduces are:

* OP_SELECT_GE
* OP_SELECT_GT
* OP_SELECT_LE
* OP_SELECT_LT

These all accept 4 operands. The first two are the basis of the comparison, then second two are the "then" and "else" case. These 4 instructions break down to some mov instructions, a new OP_CMP instruction (the low-level comparison), and a corresponding OP_CMOV_ instruction.