Ensure getinlinecache generates a block
Previously, opt_getinlinecache would return YJIT_CANT_COMPILE when the
cache was outdated. If this was being compiled through
rb_yjit_compile_iseq/gen_entry_point it would return NULL as the
jit_func.

We continue to return YJIT_CANT_COMPILE for the cases we know we can
never compile.