Reset total_calls when invalidating jit_func
When we invalidate the first block (at idx=0) on an iseq we unset the
jit_func.

Previously, if we'd previously compiled via the jit func, we would never
attempt recompilation (at least not via jit_func) because we'd exceeded
total_calls.

This commit resets total_calls to 0 when clearing jit_func.